+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## x.y.z (yyyy-mm-dd)
+
+  - More thorough instantiation inspections.
+
 ## 0.0.1 (2013-09-14)
 
   - Initial revision.